function is move to bussiness logic mean server action
