+++ src/db/data-access/communities/query.ts
@@ -545,27 +545,3 @@ export const getCommunitiesByIds = async function (communityIds: string[]) {
     throw new Error("Failed to fetch communities")
   }
 }
-
-export const ensureCommunityMembership = async (
-  communityId: string,
-  userId: string
-): Promise<void> => {
-  const communityMembers = await getCommunityUsers(communityId)
-  const communityUserIds = communityMembers.map((cu) => cu.user_id)
-
-  const isMember = communityUserIds.includes(userId)
-
-  if (!isMember) {
-    const attachCommunityUserRole = await getAndAssignViewerRoles(
-      userId,
-      "community_viewer",
-      communityId
-    )
-
-    await attachCommunityUser(
-      communityId,
-      userId,
-      attachCommunityUserRole?.viewerRole?.name
-    )
-  }
-}
