+++ src/server-actions/Channel/Channel.ts
@@ -31,11 +31,7 @@ import {
 } from "@/src/db/data-access/roles/query"
 import { GetUserPermissionsParsedAction } from "../UserRoles/UserRole"
 import { PermissionChecker } from "@/src/lib/PermissionCheker"
-import {
-  attachCommunityUser,
-  ensureCommunityMembership,
-  getCommunityUsers
-} from "@/src/db/data-access/communities/query"
+import { ensureCommunityMembership } from "../Community/Community"
 
 export const CreateChannelAction = CreateServerAction(
   true,
