+++ src/server-actions/Community/Community.ts
@@ -312,3 +312,28 @@ export const GetFeaturedCommunitiesAction = CreateServerAction(
     }
   }
 )
+
+// this function will check if the user is a member of the community
+export const ensureCommunityMembership = async (
+  communityId: string,
+  userId: string
+): Promise<void> => {
+  const communityMembers = await getCommunityUsers(communityId)
+  const communityUserIds = communityMembers.map((cu) => cu.user_id)
+
+  const isMember = communityUserIds.includes(userId)
+
+  if (!isMember) {
+    const attachCommunityUserRole = await getAndAssignViewerRoles(
+      userId,
+      "community_viewer",
+      communityId
+    )
+
+    await attachCommunityUser(
+      communityId,
+      userId,
+      attachCommunityUserRole?.viewerRole?.name
+    )
+  }
+}
